modified:   .gitignore deleted:    readme.txt modified:   src/main.py readme.md src/data/

+++ src/main.py
@@ -13,7 +13,11 @@ imgURL = ky_etrees.get_one_data_by_xpath("//link[@id=\"preloadBg\"]/@href")
 # 下载图片
 if not os.path.exists("./data/"):
     os.mkdir("./data/")
-save_path = "./data/" + time.strftime('%Y-%m-%d_%H-%M-%S') + "_" + title + ".jpg"
-img_response = ky_requests.get(imgURL)
-with open(save_path, "wb") as f:
-    f.write(img_response.content)
+save_path = "./data/" + time.strftime('%Y-%m-%d') + "_" + title + ".jpg"
+if not os.path.exists(save_path):
+    img_response = ky_requests.get(imgURL)
+    with open(save_path, "wb") as f:
+        f.write(img_response.content)
+
+# 提示
+print("程序执行结束")
